extract lazy creation tests to separate file
